split test case
